add update user

+++ src/main/java/com/example/demo/service/UserService.java
@@ -7,7 +7,9 @@ import com.example.demo.model.User;
 import com.example.demo.repository.UserRepo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
+import java.util.Objects;
 import java.util.Optional;
 
 @Service
@@ -22,6 +24,7 @@ public class UserService {
         if (existingUser != null) {
             throw new UserAlreadyExistException("пользователь " + user.getUsername() + " уже существуюет");
         }
+        user.setDob(user.getAge());
         return userRepo.save(user);
     }
 
@@ -38,14 +41,31 @@ public class UserService {
         throw new UserNotFoundException("пользователь c Id: " + id + " не найден");
     }
 
-    ;
+ /*
+  * @Transactional - транзакция, если ошибка то откат и возвращает не кастомную ошибку,
+  * а ошибку о транзакции в целом
+  */
+@Transactional
+    public UserEntity updateUser(Long id, Long age) throws IllegalStateException {
+        /*
+         * В контексте Transactional выводит в консоль ошибку и прерывается
+         */
+        UserEntity user = userRepo.findById(id).orElseThrow(() -> new IllegalStateException(
+                "user with id= " + id + "does not exist")
+        );
+        if (age != null && !Objects.equals(user.getAge(), age)) {
+            user.setAge(age);
+        }
+        userRepo.save(user);
+        return user;
+    }
 
     public Long deleteUser(Long id) throws UserNotFoundException {
-        System.out.println("findUser(id);");
-        System.out.println(findUser(id));
+        boolean exist = userRepo.existsById(id);
+        if (!exist) {
+            throw new IllegalStateException("user with id= " + id + "does not exist");
+        }
         userRepo.deleteById(id);
         return id;
-        }
-
-
+    }
 }
